# swagger-export.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/countries/{id}/cities": {
      "get": {
        "operationId": "CountriesController_findCitiesByCountry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Country ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cities in the country",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "name": "New York",
                    "countryId": 1,
                    "createdAt": "2025-01-01T00:00:00.000Z",
                    "updatedAt": "2025-01-01T00:00:00.000Z"
                  }
                ]
              }
            }
          }
        },
        "summary": "Get cities by country",
        "tags": [
          "🗺️ Countries"
        ]
      }
    },
...
    "/challenges": {
      "post": {
        "description": "Create a programming/design challenge with TechCoin rewards. Requires TechCoin for reward pool.",
        "operationId": "ChallengesController_createChallenge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Challenge created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 10,
                    "title": "Code Master Challenge",
                    "description": "Build a web app using React and Node.js",
                    "reward": 100,
                    "entryFee": 25,
                    "type": "VOTE",
                    "status": "ACTIVE",
                    "startDate": "2025-07-20T00:00:00Z",
                    "endDate": "2025-07-27T23:59:59Z",
                    "createdBy": {
                      "id": 123,
                      "username": "johndoe",
                      "avatar": null
                    },
                    "participantCount": 0,
                    "createdAt": "2025-07-16T14:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Insufficient TechCoin for reward pool"
          },
          "401": {
            "description": "Unauthorized - JWT token required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new challenge",
        "tags": [
          "🏆 Challenges & Competitions"
        ]
      },
      "get": {
        "description": "Retrieve paginated list of all challenges with filtering options",
        "operationId": "ChallengesController_getAllChallenges",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 10)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by challenge status",
            "schema": {
              "enum": [
                "ACTIVE",
                "COMPLETED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by challenge type",
            "schema": {
              "enum": [
                "VOTE",
                "JUDGE"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Challenges retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "challenges": [
                      {
                        "id": 10,
                        "title": "Code Master Challenge",
                        "description": "Build a web app using React and Node.js",
                        "reward": 100,
                        "entryFee": 25,
                        "type": "VOTE",
                        "status": "ACTIVE",
                        "participantCount": 5,
                        "startDate": "2025-07-20T00:00:00Z",
                        "endDate": "2025-07-27T23:59:59Z",
                        "createdBy": {
                          "id": 123,
                          "username": "johndoe"
                        }
                      }
                    ],
                    "pagination": {
                      "page": 1,
                      "limit": 10,
                      "total": 25,
                      "totalPages": 3
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all challenges",
        "tags": [
          "🏆 Challenges & Competitions"
        ]
      }
    },
    "/challenges/my-created": {
      "get": {
        "operationId": "ChallengesController_getMyCreatedChallenges",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User created challenges retrieved"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get challenges created by current user",
        "tags": [
          "🏆 Challenges & Competitions"
        ]
      }
    },
    "/challenges/my-participated": {
      "get": {
        "operationId": "ChallengesController_getMyParticipatedChallenges",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User participated challenges retrieved"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get challenges user participated in",
        "tags": [
          "🏆 Challenges & Competitions"
        ]
      }
    },
    "/challenges/{id}": {
      "get": {
        "operationId": "ChallengesController_getChallengeById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Challenge retrieved successfully"
          },
          "404": {
            "description": "Challenge not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get challenge by ID",
        "tags": [
          "🏆 Challenges & Competitions"
        ]
      }
    },
    "/challenges/{id}/join": {
      "post": {
        "operationId": "ChallengesController_joinChallenge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Joined challenge successfully"
          },
          "400": {
            "description": "Cannot join challenge"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join a challenge",
        "tags": [
          "🏆 Challenges & Competitions"
        ]
      }
    },
    "/challenges/{id}/submit": {
      "post": {
        "operationId": "ChallengesController_submitSolution",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitChallengeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solution submitted successfully"
          },
          "404": {
            "description": "Not a participant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit solution to challenge",
        "tags": [
          "🏆 Challenges & Competitions"
        ]
      }
    },
    "/challenges/{id}/vote/{participantId}": {
      "post": {
        "operationId": "ChallengesController_voteForParticipant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "participantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vote recorded successfully"
          },
          "403": {
            "description": "Not authorized to vote"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vote for a participant",
        "tags": [
          "🏆 Challenges & Competitions"
        ]
      }
    },
    "/challenges/{id}/close": {
      "post": {
        "operationId": "ChallengesController_closeChallenge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Challenge closed successfully"
          },
          "403": {
            "description": "Only creator can close challenge"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close challenge and determine winner",
        "tags": [
          "🏆 Challenges & Competitions"
        ]
      }
    },
...
  },
  "info": {
    "title": "TechVerse Café API",
    "description": "Backend API for TechVerse Café MVP - A tech community platform with digital economy\n\n## 🚀 New Features in MVP\n- � **Digital Wallet & TechCoin System**: Earn, buy, and spend TechCoin\n- � **Challenges & Competitions**: Community challenges with rewards\n- �️ **Enhanced Projects System**: Task management with payments\n- 🌍 **Multi-language Support**: Localized content and UI\n- ☕ **Café Communities**: Topic-based discussion groups\n\n## Core Features\n- � JWT Authentication & Role-based Access Control\n- 📝 Content Management (Articles, Posts)\n- 🔖 Bookmarking System\n- 👥 User Profiles & Social Features\n- 🤖 AI Integration for content generation\n\n## Authentication\nMost endpoints require authentication using JWT tokens. Include the token in the `X-HTTP-TOKEN` header.\n\n## Role-Based Access Control\nThe platform uses a two-tier role system:\n- **Global Roles**: ADMIN, SUPERVISOR, EDITOR, MEMBER\n- **Café Roles**: BARISTA, THINKER, JOURNALIST, MEMBER\n\n## Getting Started\n1. Register a new user account\n2. Login to receive JWT token\n3. Use the token in `X-HTTP-TOKEN` header for authenticated requests\n4. Create your wallet and start earning TechCoin\n5. Join challenges and participate in projects\n\n## Download API Specification\n- JSON Format: [/api/swagger.json](/api/swagger.json)\n- YAML Format: [/api/swagger.yaml](/api/swagger.yaml)\n      ",
    "version": "2.0.0",
...
    },
...
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "X-HTTP-TOKEN": {
        "type": "apiKey",
        "in": "header",
        "name": "X-HTTP-TOKEN",
        "description": "JWT token for authentication. Format: Bearer <token>"
      }
    },
    "schemas": {
...
      "ChallengeType": {
        "type": "string",
        "enum": [
          "VOTE",
          "JUDGE"
        ],
        "description": "How the winner will be determined"
      },
      "CreateChallengeDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Challenge title",
            "example": "Code Master Challenge",
            "minLength": 5,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Detailed challenge description and requirements",
            "example": "Build a responsive web application using React and Node.js. Must include user authentication, CRUD operations, and responsive design.",
            "minLength": 20,
            "maxLength": 2000
          },
          "reward": {
            "type": "number",
            "description": "TechCoin reward for the winner",
            "example": 100,
            "minimum": 10,
            "maximum": 10000
          },
          "entryFee": {
            "type": "number",
            "description": "TechCoin entry fee required to participate",
            "example": 25,
            "minimum": 5,
            "maximum": 1000
          },
          "type": {
            "description": "How the winner will be determined",
            "example": "VOTE",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChallengeType"
              }
            ]
          },
          "startDate": {
            "type": "string",
            "description": "Challenge start date and time (ISO 8601)",
            "example": "2025-07-20T00:00:00Z"
          },
          "endDate": {
            "type": "string",
            "description": "Challenge end date and time (ISO 8601)",
            "example": "2025-07-27T23:59:59Z"
          },
          "requirements": {
            "type": "string",
            "description": "Additional requirements or rules",
            "example": "Code must be submitted via GitHub repository link",
            "maxLength": 1000
          },
          "tags": {
            "description": "Tags for challenge categorization",
            "example": [
              "javascript",
              "react",
              "nodejs",
              "fullstack"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "description",
          "reward",
          "entryFee",
          "type",
          "startDate",
          "endDate"
        ]
      },
      "JoinChallengeDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Optional message or approach description when joining",
            "example": "I plan to use React with TypeScript and implement real-time features using Socket.io",
            "maxLength": 500
          },
          "portfolioUrl": {
            "type": "string",
            "description": "GitHub or portfolio URL (optional)",
            "example": "https://github.com/username",
            "maxLength": 255
          }
        }
      },
      "SubmitChallengeDto": {
        "type": "object",
        "properties": {
          "submissionUrl": {
            "type": "string",
            "description": "URL to the project submission (GitHub, live demo, etc.)",
            "example": "https://github.com/username/challenge-project",
            "format": "url"
          },
          "demoUrl": {
            "type": "string",
            "description": "Live demo URL (optional)",
            "example": "https://my-challenge-app.vercel.app",
            "format": "url"
          },
          "description": {
            "type": "string",
            "description": "Description of the submission and approach taken",
            "example": "Built a full-stack e-commerce app with React, Node.js, and PostgreSQL. Features include user auth, product catalog, shopping cart, and payment integration.",
            "maxLength": 2000
          },
          "technologies": {
            "description": "Technologies used in the project",
            "example": [
              "React",
              "Node.js",
              "PostgreSQL",
              "Stripe",
              "TailwindCSS"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "description": "Additional notes for judges/voters",
            "example": "Spent extra time on mobile responsiveness and accessibility features.",
            "maxLength": 500
          }
        },
        "required": [
          "submissionUrl",
          "description"
        ]
      },
...
    }
  }
}